Backend: Add more exceptions 3. Remove redundant code.

+++ src/main/java/com/openu/project/business/service/ImageService.java
@@ -1,5 +1,6 @@
 package com.openu.project.business.service;
 
+import com.openu.project.exception.exceptionsList.FileAlreadyExist;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -9,16 +10,28 @@ import java.nio.file.Paths;
 @Service
 public class ImageService {
     String imageDbPath = "C:/db/pic_matok/";
-    String imageNotFoundPath =  imageDbPath + "/not_found.png";
     public byte[] getPhotoService(String name, String category) throws IOException {
         String image_path = imageDbPath + category + '/' + name;
-        //System.out.println(image_path);
+
         byte[] image;
-        try{
-            image = Files.readAllBytes(Paths.get(image_path));
-        } catch (Exception e)
+        if (Files.exists(Paths.get(image_path)))
         {
-            image = Files.readAllBytes(Paths.get(imageNotFoundPath));
+            try {
+                image = Files.readAllBytes(Paths.get(image_path));
+            } catch (Exception e) {
+                System.out.println(e.getMessage());
+                throw e;
+            }
+        }
+        else
+        {
+            try {
+                String imageNotFoundPath =  imageDbPath + "/not_found.png";
+                image = Files.readAllBytes(Paths.get(imageNotFoundPath));
+            } catch (Exception e) {
+                System.out.println(e.getMessage());
+                throw e;
+            }
         }
         return image;
     }
@@ -26,7 +39,8 @@ public class ImageService {
 
     public boolean savePhotoService(byte[] image, String imageName, String category) throws IOException {
         String image_path = imageDbPath + category + '/' + imageName;
-        //System.out.println(image_path);
+        if (Files.exists(Paths.get(image_path))) throw new FileAlreadyExist();
+
         try {
             Files.write(Paths.get(image_path), image);
         } catch (IOException e) {
